refactor(console): split session test suite
